+++ src/set/gnunet-service-set_union.c
@@ -931,7 +931,7 @@ handle_union_p2p_strata_estimator (void *cls,
     LOG (GNUNET_ERROR_TYPE_DEBUG,
          "Deciding to go for full set transmission (diff=%d, own set=%llu)\n",
          diff,
-         op->state->initial_size);
+         (unsigned long long) op->state->initial_size);
     GNUNET_STATISTICS_update (_GSS_statistics,
                               "# of full sends",
                               1,
